add TextParagraph
